+++ backend/matching-service/.env.example
@@ -1,3 +1,4 @@
 ENV=PROD
+SERVICE_PORT=
 DB_CLOUD_URI=
 DB_LOCAL_URI=
