+++ backend/question-service/.env.example
@@ -1,0 +1,4 @@
+ENV=PROD
+SERVICE_PORT=
+DB_CLOUD_URI=
+DB_LOCAL_URI=
