+++ backend/question-service/.gitignore
@@ -1,0 +1,14 @@
+# dependencies
+/node_modules
+
+# production
+/build
+
+# misc
+.DS_Store
+
+npm-debug.log*
+*.log
+*.gz
+secret.*
+.env
