--- matching-service/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "matching-service",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "dev": "nodemon index.js",
-    "start": "node index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "nodemon": "^2.0.19"
-  },
-  "dependencies": {
-    "dotenv": "^16.0.2",
-    "express": "^4.18.1",
-    "express-validator": "^6.14.2",
-    "mongoose": "^6.6.0",
-    "sequelize": "^6.21.3",
-    "socket.io": "^4.5.2",
-    "sqlite3": "^5.0.10",
-    "uuid": "^8.3.2"
-  }
-}
