--- user-service/model/user-model.js
@@ -1,15 +1,0 @@
-import mongoose from 'mongoose';
-var Schema = mongoose.Schema
-let UserModelSchema = new Schema({
-    username: {
-        type: String,
-        required: true,
-        unique: true,
-    },
-    password: {
-        type: String,
-        required: true,
-    }
-})
-
-export default mongoose.model('UserModel', UserModelSchema)
